proof of concept inverted inheritance HOC
