+++ src/components/victory-bar/victory-bar.js
@@ -1,12 +1,12 @@
-import { assign, defaults, isFunction, partialRight } from "lodash";
+import { partialRight } from "lodash";
 import React, { PropTypes } from "react";
 import BarHelpers from "./helper-methods";
+import VictoryBase from "../victory-base/victory-base";
+import cartesianProps from "../victory-base/cartesian-props";
+import commonProps from "../victory-base/common-props";
 import Data from "../../helpers/data";
 import Domain from "../../helpers/domain";
-import {
-  PropTypes as CustomPropTypes, Helpers, Events, VictoryTransition, VictoryLabel,
-  VictoryContainer, VictoryTheme, Bar, ClipPath
-} from "victory-core";
+import { VictoryLabel, VictoryContainer, VictoryTheme, Bar } from "victory-core";
 
 const fallbackProps = {
   width: 450,
@@ -21,7 +21,7 @@ const defaultData = [
   {x: 4, y: 4}
 ];
 
-export default class VictoryBar extends React.Component {
+class VictoryBar extends React.Component {
   static displayName = "VictoryBar";
 
   static role = "bar";
@@ -44,309 +44,9 @@ export default class VictoryBar extends React.Component {
   };
 
   static propTypes = {
-    /**
-     * The animate prop specifies props for VictoryAnimation to use. The animate prop should
-     * also be used to specify enter and exit transition configurations with the `onExit`
-     * and `onEnter` namespaces respectively.
-     * @examples {duration: 500, onEnd: () => {}, onEnter: {duration: 500, before: () => ({y: 0})})}
-     */
-    animate: PropTypes.object,
-    /**
-     * The categories prop specifies how categorical data for a chart should be ordered.
-     * This prop should be given as an array of string values, or an object with
-     * these arrays of values specified for x and y. If this prop is not set,
-     * categorical data will be plotted in the order it was given in the data array
-     * @examples ["dogs", "cats", "mice"]
-     */
-    categories: PropTypes.oneOfType([
-      PropTypes.arrayOf(PropTypes.string),
-      PropTypes.shape({
-        x: PropTypes.arrayOf(PropTypes.string),
-        y: PropTypes.arrayOf(PropTypes.string)
-      })
-    ]),
-    /**
-     * The data prop specifies the data to be plotted. Data should be in the form of an array
-     * of data points. Each data point may be any format you wish
-     * (depending on the `x` and `y` accessor props), but by default, an object
-     * with x and y properties is expected.
-     * @examples [{x: 1, y: 2}, {x: 2, y: 3}], [[1, 2], [2, 3]],
-     * [[{x: "a", y: 1}, {x: "b", y: 2}], [{x: "a", y: 2}, {x: "b", y: 3}]]
-     */
-    data: PropTypes.array,
-    /**
-     * The dataComponent prop takes an entire component which will be used to create bars for
-     * each datum in the chart. The new element created from the passed dataComponent will be
-     * provided with the following properties calculated by VictoryBar: datum, index, scale,
-     * style, events, horizontal (boolean), x, y, and y0. Any of these props may be overridden
-     * by passing in props to the supplied component, or modified or ignored within the custom
-     * component itself. If a dataComponent is not provided, VictoryBar will use its default
-     * Bar component.
-     */
-    dataComponent: PropTypes.element,
-    /**
-     * The domainPadding prop specifies a number of pixels of padding to add to the
-     * beginning and end of a domain. This prop is useful for explicitly spacing ticks farther
-     * from the origin to prevent crowding. This prop should be given as an object with
-     * numbers specified for x and y.
-     */
-    domainPadding: PropTypes.oneOfType([
-      PropTypes.shape({
-        x: PropTypes.oneOfType([
-          PropTypes.number,
-          CustomPropTypes.domain
-        ]),
-        y: PropTypes.oneOfType([
-          PropTypes.number,
-          CustomPropTypes.domain
-        ])
-      }),
-      PropTypes.number
-    ]),
-    /**
-     * The domain prop describes the range of values your bar chart will cover. This prop can be
-     * given as a array of the minimum and maximum expected values for your bar chart,
-     * or as an object that specifies separate arrays for x and y.
-     * If this prop is not provided, a domain will be calculated from data, or other
-     * available information.
-     * @examples [-1, 1], {x: [0, 100], y: [0, 1]}
-     */
-    domain: PropTypes.oneOfType([
-      CustomPropTypes.domain,
-      PropTypes.shape({
-        x: CustomPropTypes.domain,
-        y: CustomPropTypes.domain
-      })
-    ]),
-    /**
-     * The event prop take an array of event objects. Event objects are composed of
-     * a target, an eventKey, and eventHandlers. Targets may be any valid style namespace
-     * for a given component, so "data" and "labels" are all valid targets for VictoryBar events.
-     * The eventKey may optionally be used to select a single element by index rather than an entire
-     * set. The eventHandlers object should be given as an object whose keys are standard
-     * event names (i.e. onClick) and whose values are event callbacks. The return value
-     * of an event handler is used to modify elemnts. The return value should be given
-     * as an object or an array of objects with optional target and eventKey keys,
-     * and a mutation key whose value is a function. The target and eventKey keys
-     * will default to those corresponding to the element the event handler was attached to.
-     * The mutation function will be called with the calculated props for the individual selected
-     * element (i.e. a single bar), and the object returned from the mutation function
-     * will override the props of the selected element via object assignment.
-     * @examples
-     * events={[
-     *   {
-     *     target: "data",
-     *     eventKey: "thisOne",
-     *     eventHandlers: {
-     *       onClick: () => {
-     *         return [
-     *            {
-     *              eventKey: "theOtherOne",
-     *              mutation: (props) => {
-     *                return {style: merge({}, props.style, {fill: "orange"})};
-     *              }
-     *            }, {
-     *              eventKey: "theOtherOne",
-     *              target: "labels",
-     *              mutation: () => {
-     *                return {text: "hey"};
-     *              }
-     *            }
-     *          ];
-     *       }
-     *     }
-     *   }
-     * ]}
-     *}}
-     */
-    events: PropTypes.arrayOf(PropTypes.shape({
-      target: PropTypes.oneOf(["data", "labels", "parent"]),
-      eventKey: PropTypes.oneOfType([
-        PropTypes.array,
-        CustomPropTypes.allOfType([CustomPropTypes.integer, CustomPropTypes.nonNegative]),
-        PropTypes.string
-      ]),
-      eventHandlers: PropTypes.object
-    })),
-    /**
-     * The name prop is used to reference a component instance when defining shared events.
-     */
-    name: PropTypes.string,
-    /**
-     * Similar to data accessor props `x` and `y`, this prop may be used to functionally
-     * assign eventKeys to data
-     */
-    eventKey: PropTypes.oneOfType([
-      PropTypes.func,
-      CustomPropTypes.allOfType([CustomPropTypes.integer, CustomPropTypes.nonNegative]),
-      PropTypes.string
-    ]),
-    /**
-     * This prop is used to coordinate events between VictoryArea and other Victory
-     * Components via VictorySharedEvents. This prop should not be set manually.
-     */
-    sharedEvents: PropTypes.shape({
-      events: PropTypes.array,
-      getEventState: PropTypes.func
-    }),
-    /**
-     * The height props specifies the height the svg viewBox of the chart container.
-     * This value should be given as a number of pixels
-     */
-    height: CustomPropTypes.nonNegative,
-    /**
-     * The horizontal prop determines whether the bars will be laid vertically or
-     * horizontally. The bars will be vertical if this prop is false or unspecified,
-     * or horizontal if the prop is set to true.
-     */
-    horizontal: PropTypes.bool,
-    /**
-     * The labels prop defines labels that will appear above each bar in your bar chart.
-     * This prop should be given as an array of values or as a function of data.
-     * If given as an array, the number of elements in the array should be equal to
-     * the length of the data array. Labels may also be added directly to the data object
-     * like data={[{x: 1, y: 1, label: "first"}]}.
-     * @examples: ["spring", "summer", "fall", "winter"], (datum) => datum.title
-     */
-    labels: PropTypes.oneOfType([
-      PropTypes.func,
-      PropTypes.array
-    ]),
-    /**
-     * The labelComponent prop takes in an entire label component which will be used
-     * to create labels for each bar in the bar chart. The new element created from
-     * the passed labelComponent will be supplied with the following properties:
-     * x, y, y0, index, datum, verticalAnchor, textAnchor, angle, style, text, and events.
-     * Any of these props may be overridden by passing in props to the supplied component,
-     * or modified or ignored within the custom component itself. If labelComponent is omitted,
-     * a new VictoryLabel will be created with props described above.
-     */
-    labelComponent: PropTypes.element,
-    /**
-     * The padding props specifies the amount of padding in number of pixels between
-     * the edge of the chart and any rendered child components. This prop can be given
-     * as a number or as an object with padding specified for top, bottom, left
-     * and right.
-     */
-    padding: PropTypes.oneOfType([
-      PropTypes.number,
-      PropTypes.shape({
-        top: PropTypes.number,
-        bottom: PropTypes.number,
-        left: PropTypes.number,
-        right: PropTypes.number
-      })
-    ]),
-    /**
-     * The scale prop determines which scales your chart should use. This prop can be
-     * given as a string specifying a supported scale ("linear", "time", "log", "sqrt"),
-     * as a d3 scale function, or as an object with scales specified for x and y
-     * @exampes d3Scale.time(), {x: "linear", y: "log"}
-     */
-    scale: PropTypes.oneOfType([
-      CustomPropTypes.scale,
-      PropTypes.shape({
-        x: CustomPropTypes.scale,
-        y: CustomPropTypes.scale
-      })
-    ]),
-    /**
-     * The standalone prop determines whether the component will render a standalone svg
-     * or a <g> tag that will be included in an external svg. Set standalone to false to
-     * compose VictoryBar with other components within an enclosing <svg> tag.
-     */
-    standalone: PropTypes.bool,
-    /**
-     * The style prop specifies styles for your VictoryBar. Any valid inline style properties
-     * will be applied. Height, width, and padding should be specified via the height,
-     * width, and padding props, as they are used to calculate the alignment of
-     * components within chart. In addition to normal style properties, angle and verticalAnchor
-     * may also be specified via the labels object, and they will be passed as props to
-     * VictoryLabel, or any custom labelComponent.
-     * @examples {data: {fill: "red", width: 8}, labels: {fontSize: 12}}
-     */
-    style: PropTypes.shape({
-      parent: PropTypes.object,
-      data: PropTypes.object,
-      labels: PropTypes.object
-    }),
-    /**
-     * The width props specifies the width of the svg viewBox of the chart container
-     * This value should be given as a number of pixels
-     */
-    width: CustomPropTypes.nonNegative,
-    /**
-     * The x prop specifies how to access the X value of each data point.
-     * If given as a function, it will be run on each data point, and returned value will be used.
-     * If given as an integer, it will be used as an array index for array-type data points.
-     * If given as a string, it will be used as a property key for object-type data points.
-     * If given as an array of strings, or a string containing dots or brackets,
-     * it will be used as a nested object property path (for details see Lodash docs for _.get).
-     * If `null` or `undefined`, the data value will be used as is (identity function/pass-through).
-     * @examples 0, 'x', 'x.value.nested.1.thing', 'x[2].also.nested', null, d => Math.sin(d)
-     */
-    x: PropTypes.oneOfType([
-      PropTypes.func,
-      CustomPropTypes.allOfType([CustomPropTypes.integer, CustomPropTypes.nonNegative]),
-      PropTypes.string,
-      PropTypes.arrayOf(PropTypes.string)
-    ]),
-    /**
-     * The y prop specifies how to access the Y value of each data point.
-     * If given as a function, it will be run on each data point, and returned value will be used.
-     * If given as an integer, it will be used as an array index for array-type data points.
-     * If given as a string, it will be used as a property key for object-type data points.
-     * If given as an array of strings, or a string containing dots or brackets,
-     * it will be used as a nested object property path (for details see Lodash docs for _.get).
-     * If `null` or `undefined`, the data value will be used as is (identity function/pass-through).
-     * @examples 0, 'y', 'y.value.nested.1.thing', 'y[2].also.nested', null, d => Math.sin(d)
-     */
-    y: PropTypes.oneOfType([
-      PropTypes.func,
-      CustomPropTypes.allOfType([CustomPropTypes.integer, CustomPropTypes.nonNegative]),
-      PropTypes.string,
-      PropTypes.arrayOf(PropTypes.string)
-    ]),
-    /**
-     * The containerComponent prop takes an entire component which will be used to
-     * create a container element for standalone charts.
-     * The new element created from the passed containerComponent wil be provided with
-     * these props from VictoryBar: height, width, children
-     * (the chart itself) and style. Props that are not provided by the
-     * child chart component include title and desc, both of which
-     * are intended to add accessibility to Victory components. The more descriptive these props
-     * are, the more accessible your data will be for people using screen readers.
-     * Any of these props may be overridden by passing in props to the supplied component,
-     * or modified or ignored within the custom component itself. If a dataComponent is
-     * not provided, VictoryBar will use the default VictoryContainer component.
-     * @examples <VictoryContainer title="Chart of Dog Breeds" desc="This chart shows how
-     * popular each dog breed is by percentage in Seattle." />
-     */
-    containerComponent: PropTypes.element,
-    /**
-    * The theme prop takes a style object with nested data, labels, and parent objects.
-    * You can create this object yourself, or you can use a theme provided by Victory.
-    * When using VictoryBar as a solo component, implement the theme directly on
-    * VictoryBar. If you are wrapping VictoryBar in VictoryChart, VictoryStack, or
-    * VictoryGroup, please call the theme on the outermost wrapper component instead.
-    * @examples theme={VictoryTheme.material}
-    */
-    theme: PropTypes.object,
-    /**
-     * The groupComponent prop takes an entire component which will be used to
-     * create group elements for use within container elements. This prop defaults
-     * to a <g> tag on web, and a react-native-svg <G> tag on mobile
-     */
-    groupComponent: PropTypes.element,
-    /**
-     * The clipPathComponent prop takes an entire component which will be used to
-     * create clipPath elements for use within container elements.
-     */
-    clipPathComponent: PropTypes.element,
-    /**
-     * Unique clipId for clipPath
-     */
-    clipId: PropTypes.number
+    ...cartesianProps,
+    ...commonProps,
+    horizontal: PropTypes.bool
   };
 
   static defaultProps = {
@@ -359,149 +59,20 @@ export default class VictoryBar extends React.Component {
     y: "y",
     containerComponent: <VictoryContainer/>,
     groupComponent: <g/>,
-    clipPathComponent: <ClipPath/>,
     theme: VictoryTheme.grayscale
   };
 
   static getDomain = Domain.getDomainWithZero.bind(Domain);
   static getData = Data.getData.bind(Data);
   static getBaseProps = partialRight(BarHelpers.getBaseProps.bind(BarHelpers), fallbackProps);
-
-  constructor() {
-    super();
-    this.state = {};
-    const getScopedEvents = Events.getScopedEvents.bind(this);
-    this.getEvents = partialRight(Events.getEvents.bind(this), getScopedEvents);
-    this.getEventState = Events.getEventState.bind(this);
-  }
-
-  componentWillMount() {
-    this.setupEvents(this.props);
-  }
-
-  componentWillReceiveProps(newProps) {
-    this.setupEvents(newProps);
-  }
-
-  setupEvents(props) {
-    const { sharedEvents } = props;
-    const components = ["dataComponent", "labelComponent", "groupComponent", "containerComponent"];
-    this.componentEvents = Events.getComponentEvents(props, components);
-    this.baseProps = BarHelpers.getBaseProps(props, fallbackProps);
-    this.dataKeys = Object.keys(this.baseProps).filter((key) => key !== "parent");
-    this.getSharedEventState = sharedEvents && isFunction(sharedEvents.getEventState) ?
-      sharedEvents.getEventState : () => undefined;
-    this.hasEvents = props.events || props.sharedEvents || this.componentEvents;
-  }
-
-  renderData(props) {
-    const { dataComponent, labelComponent, groupComponent, clipId } = props;
-    const { role } = VictoryBar;
-    const dataComponents = [];
-    const labelComponents = [];
-    const getComponentProps = (index, component, type) => {
-      const key = this.dataKeys[index];
-      if (this.hasEvents) {
-        const events = this.getEvents(props, type, key);
-        const componentProps = defaults(
-          {index, key: `${role}-${type}-${key}`, role: `${role}-${index}`, clipId},
-          this.getEventState(key, type),
-          this.getSharedEventState(key, type),
-          component.props,
-          this.baseProps[key][type]
-        );
-        return assign(
-          {}, componentProps, {events: Events.getPartialEvents(events, key, componentProps)}
-        );
-      }
-      return defaults(
-        {index, key: `${role}-${type}-${key}`, role: `${role}-${index}`, clipId},
-        component.props,
-        this.baseProps[key][type]
-      );
-    };
-
-    for (let index = 0, len = this.dataKeys.length; index < len; index++) {
-      const key = this.dataKeys[index];
-      const dataProps = getComponentProps(index, dataComponent, "data");
-      dataComponents[index] = React.cloneElement(dataComponent, dataProps);
-
-      if (this.baseProps[key].labels || this.hasEvents) {
-        const labelProps = getComponentProps(index, labelComponent, "labels");
-        if (labelProps && labelProps.text !== undefined && labelProps.text !== null) {
-          labelComponents[index] = React.cloneElement(labelComponent, labelProps);
-        }
-      }
-    }
-    return labelComponents.length > 0 ?
-      React.cloneElement(groupComponent, {}, ...dataComponents, ...labelComponents) :
-      dataComponents;
-  }
-
-  renderContainer(props, group) {
-    let parentProps;
-    if (this.hasEvents) {
-      const parentEvents = this.getEvents(props, "parent", "parent");
-      const baseProps = defaults(
-        {},
-        this.getEventState("parent", "parent"),
-        this.getSharedEventState("parent", "parent"),
-        props.containerComponent.props,
-        this.baseProps.parent
-      );
-      parentProps = assign(
-        {}, baseProps, {events: Events.getPartialEvents(parentEvents, "parent", baseProps)}
-      );
-    } else {
-      parentProps = defaults({}, props.containerComponent.props, this.baseProps.parent);
-    }
-
-    return React.cloneElement(props.containerComponent, parentProps, group);
-  }
-
-  renderGroup(children, props, style) {
-    const { clipPathComponent, horizontal } = props;
-    const barWidth = BarHelpers.getBarWidth(props);
-    const clipPadding = horizontal ?
-      {"top": barWidth, bottom: barWidth} : {left: barWidth, right: barWidth};
-    const clipComponent = React.cloneElement(clipPathComponent, {
-      padding: props.padding,
-      clipId: props.clipId,
-      clipWidth: (props.clipWidth || props.width),
-      clipHeight: (props.clipHeight || props.height),
-      clipPadding
-    });
-
-    return React.cloneElement(
-      props.groupComponent,
-      { role: "presentation", style: style.parent},
-      children,
-      clipComponent
-    );
-  }
-
-  render() {
-    const clipId = this.props.clipId || Math.round(Math.random() * 10000);
-    const props = Helpers.modifyProps(assign({clipId}, this.props), fallbackProps, "bar");
-    const { animate, style, standalone, theme } = props;
-    if (animate) {
-      const whitelist = [
-        "data", "domain", "height", "padding", "style", "width"
-      ];
-      return (
-        <VictoryTransition animate={animate} animationWhitelist={whitelist}>
-          {React.createElement(this.constructor, props)}
-        </VictoryTransition>
-      );
-    }
-
-    const styleObject = theme && theme.bar && theme.bar.style ? theme.bar.style : {};
-
-    const baseStyles = Helpers.getStyles(style, styleObject, "auto", "100%");
-    const group = this.renderGroup(
-      this.renderData(props), props, baseStyles
-    );
-
-    return standalone ? this.renderContainer(props, group) : group;
-  }
+  static expectedComponents = [
+    "dataComponent", "labelComponent", "groupComponent", "containerComponent"
+  ];
+  static animationWhitelist = [
+    "data", "domain", "height", "padding", "style", "width"
+  ];
+  static fallbackProps = fallbackProps;
 }
+
+export default VictoryBase(VictoryBar);
+
